requirements.txt
click==8.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.8.1
Flask==2.1.0
gunicorn==20.1.0
google-cloud-bigquery==3.4.0